+++ .eslintrc.js
@@ -257,8 +257,23 @@ module.exports = {
         './packages/eslint-plugin/src/rules/**/*.ts',
       ],
       rules: {
+        'eslint-plugin/require-meta-docs-description': [
+          'error',
+          { pattern: '^(Enforce|Require|Disallow) .+[^. ]$' },
+        ],
+
         // specifically for rules - default exports makes the tooling easier
         'import/no-default-export': 'off',
+
+        'no-restricted-syntax': [
+          'error',
+          {
+            selector:
+              'ExportDefaultDeclaration Property[key.name="create"] MemberExpression[object.name="context"][property.name="options"]',
+            message:
+              "Retrieve options from create's second parameter so that defaultOptions are applied.",
+          },
+        ],
       },
     },
     // plugin rule tests