+++ packages/ast-spec/src/element/AccessorProperty/fixtures/with-annotation-with-value/fixture.ts
@@ -1,0 +1,3 @@
+class Foo {
+  accessor prop: string = 'str';
+}