+++ packages/scope-manager/tests/fixtures/class/declaration/accessor-property-type-annotation.ts
@@ -1,0 +1,4 @@
+type T = 1;
+class A {
+  prop: T;
+}